+++ app/Services/Internal/BeagleService.php
@@ -55,6 +55,10 @@ class BeagleService
             return [];
         }
 
+        if ((bool) config('instance.discover.beagle_api') == false) {
+            return [];
+        }
+
         return Cache::remember(self::DISCOVER_CACHE_KEY, now()->addHours(6), function () {
             try {
                 $res = Http::withOptions(['allow_redirects' => false])
@@ -89,6 +93,10 @@ class BeagleService
     public static function getDiscoverPosts()
     {
         if ((bool) config_cache('federation.activitypub.enabled') == false) {
+            return [];
+        }
+
+        if ((bool) config('instance.discover.beagle_api') == false) {
             return [];
         }